+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v2.6.1]
+### Fixed
+- Broken report when the dataframe is filtered using the `--abundance_threshold`.
+- Taxonomy abundances barplot was not showing more abundant species.
+
 ## [v2.6.0]
 ### Fixed
 - Broken plots caused by single quotes in NCBI taxon names.